Add files via upload
Please note that the majority of the notebook's content was not my work. As the README file states, the additional content includes mostly documentation and some additional coding. The added coding component includes the use of LightGBM and CatBoost regressors.

Also note that the documentation is in Korean. Contact me if you want English translation. Thank you.